move content of run _qt to old run and keep only one
